requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
absl-py==0.13.0
cachetools==4.2.2; python_version ~= '3.5'
certifi==2021.5.30
charset-normalizer==2.0.2; python_version >= '3'
click==8.0.1; python_version >= '3.6'
cycler==0.10.0
flask-cors==3.0.10
flask==1.0.3
google-auth-oauthlib==0.4.4; python_version >= '3.6'
google-auth==1.33.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
grpcio==1.38.1
idna==3.2; python_version >= '3'
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.1; python_version >= '3.6'
kiwisolver==1.3.1; python_version >= '3.6'
markdown==3.3.4; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
matplotlib==3.4.2
numpy==1.21.0
oauthlib==3.1.1; python_version >= '3.6'
opencv-python==4.5.3.56
pandas==1.3.0
pillow==8.3.1
protobuf==3.17.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.1
pyyaml==5.4.1
requests-oauthlib==1.3.0
requests==2.26.0
rsa==4.7.2; python_version >= '3.6'
scipy==1.7.0
seaborn==0.11.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tensorboard-data-server==0.6.1; python_version >= '3.6'
tensorboard-plugin-wit==1.8.0
tensorboard==2.5.0
thop==0.0.31.post2005241907
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==2.0.1; python_version >= '3.6'
wheel==0.36.2; python_version >= '3'
